Tag Libraries and CSRF
Added Authorize tags to display options to specif users
like
create account for anonymous users
add new books for admin

Enabled CSRF protection so that malicious user cannot change the state in the system using HTTP Post.
To achieve this we have to add
<input type="hidden" name="_csrf" value="1d630d07-7feb-4c1b-acbd-d980d5dfd50c" />
where the token is generated by the server.

To add this csrf token automatically we have to use either
1. spring form tag or
2. <sec:csrfInput />